+++ ItemExtensions/Additions/Clumps/Extension.cs
@@ -100,7 +100,7 @@ public static class ExtensionClump
         if (ModEntry.BigClumps.TryGetValue(id, out var resource) == false)
             return false;
 
-        if(clump.health.Value <= 0.0)
+        if(clump.health.Value < 0.0)
         {
             return false;
         }